** Wrong Folder
git-svn-id: http://subversion.assembla.com/svn/sienna-rift@74 314f1d10-98f6-48e5-a4d7-766c1cb0bb34

+++ Shared/Database/Characters/Characters/Character_Info.cs
@@ -13,7 +13,7 @@ public class Character_Position : DataObject
     public int Id;
 
     [DataElement()]
-    public int MapId;
+    public long MapId;
 
     [DataElement()]
     public float X;
